Refactor some stuff out of test.py into dag.py
