+++ src/test.py
@@ -3,6 +3,7 @@ import gc
 import os
 import time
 import traceback
+import types
 
 import numpy as np
 import torch
@@ -23,6 +24,8 @@ import csv
 
 import logging
 import shutil
+
+from dag import *
 
 def get_logger(save_path):
     logger = logging.getLogger('ENAS')
@@ -53,45 +56,20 @@ class Flatten(nn.Module):
     def forward(self, input):
         return input.view(input.size(0), -1)
 
-def start():
-    return time.clock()
+class Timer:
+    def __init__(self, debugText="", print_funct: types.FunctionType = print):
+        self.debugText = debugText
+        self.print_funct = print_funct
 
-def stop(startTime, debugTest="", logger=None):
-    diff = time.clock() - startTime
-    message = f"{debugTest} Diff={diff}"
-    if logger:
-        logger.info(message)
-    else:
-        print(message)
-    return diff
+    def __enter__(self):
+        self.start = time.clock()
+        return self
 
-def trim_dag_one_way(dag: list, start_nodes: set, forward = True):
-    start_nodes = set(start_nodes)
-    if forward:
-        start_index, end_index = 0, 1
-    else:
-        start_index, end_index = 1, 0
-    new_dag = []
-    while start_nodes:
-        current = start_nodes.pop()
-        connections = list(conn for conn in dag if conn[start_index] == current)
-        for conn in connections:
-            start_nodes.add(conn[end_index])
-        new_dag.extend(connections)
-    return new_dag
-
-
-def trim_dag(dag: list, start_nodes: set, output_node=None):
-    new_dag = trim_dag_one_way(dag, start_nodes, True)
-
-    if output_node and not any(conn[1] == output_node for conn in new_dag):
-        raise RuntimeError("Invalid dag")
-
-    if output_node:
-        new_dag = trim_dag_one_way(new_dag, {output_node}, False)
-
-    new_dag.sort(key = lambda x: (x[0], x[1]))
-    return new_dag
+    def __exit__(self, *args):
+        self.end = time.clock()
+        self.interval = self.end - self.start
+        message = f"{self.debugText} Diff={self.interval}"
+        self.print_funct(message)
 
 def main():
     load_path = None
@@ -108,16 +86,15 @@ def main():
     device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
     def create_cnn():
-        sT = start()
-        cnn = CNN(args, input_channels=3, height=32, width=32, output_classes=10, gpu=device)
-        cnn.train()
-        stop(sT, "CNN", logger)
+        with Timer("CNN construct", logger.debug):
+            cnn = CNN(args, input_channels=3, height=32, width=32, output_classes=10, gpu=device)
+            cnn.train()
 
-        sT = start()
-        dropout_opt = DropoutSGD([cnn.dag_variables, cnn.reducing_dag_variables], connections=cnn.all_connections,
-                                 lr=8*25*20, weight_decay=0)
-        cnn_optimizer = AdamShared(cnn.parameters(), lr=0.00001, weight_decay=1e-5)
-        stop(sT, "optimizer", logger)
+
+        with Timer("Optimizer Construct", logger.debug):
+            dropout_opt = DropoutSGD([cnn.dag_variables, cnn.reducing_dag_variables], connections=cnn.all_connections,
+                                     lr=8*25*20, weight_decay=0)
+            cnn_optimizer = AdamShared(cnn.parameters(), lr=0.00001, weight_decay=1e-5)
 
         return cnn, cnn_optimizer, dropout_opt
 
@@ -140,104 +117,6 @@ def main():
 
     print(parent_counts)
 
-    def sample(probs):
-        random_nums = np.random.rand(len(probs))
-        return np.squeeze(np.argwhere(probs >= random_nums))
-
-
-    def sample_dags(dag_probs, all_connections):
-        dag_idx = sample(dag_probs)
-        # reduction_dag_idx = sample(reduction_dag_probs)
-        if dag_idx.size > 0:
-            dag = list(all_connections[i] for i in np.nditer(dag_idx))
-        else:
-            dag = []
-
-        dag.sort()
-        return dag
-
-    def sample_valid_dag(dag_probs, reduction_dag_probs, all_connections):
-        while True:
-            dag, reduction_dag = sample_dags(dag_probs, all_connections), sample_dags(reduction_dag_probs, all_connections)
-            try:
-                dag = trim_dag(dag, {0,1}, 1 + args.num_blocks)
-                reduction_dag = trim_dag(reduction_dag, {0,1}, 1 + args.num_blocks)
-                return dag, reduction_dag
-            except RuntimeError:
-                print(f"Invalid Dags {dag}, {reduction_dag}")
-                print(f"probs: {dag_probs}, {reduction_dag_probs}")
-                continue
-
-    def fix_dag_one_way(dag: list, start_nodes: set, forward=True):
-        new_dag = list(dag)
-        start_nodes = set(start_nodes)
-        if forward:
-            start_index, end_index = 0, 1
-        else:
-            start_index, end_index = 1, 0
-
-        dag = list(dag)
-        dag.sort(key=lambda x: (x[start_index], x[end_index]))
-        if not forward:
-            dag.reverse()
-
-        # new_dag = []
-        for connection in dag:
-            start = connection[start_index]
-            end = connection[end_index]
-
-
-            if start not in start_nodes:
-                if forward:
-                    begin = max(node for node in start_nodes if node < start)
-                else:
-                    begin = min(node for node in start_nodes if node > start)
-                new_conn = [None, None, None]
-                new_conn[start_index] = begin
-                new_conn[end_index] = start
-                new_conn[2] = shared_cnn.identity.__name__
-                new_dag.append(tuple(new_conn))
-                start_nodes.add(start)
-
-            start_nodes.add(end)
-
-        new_dag.sort(key=lambda x: (x[0], x[1]))
-        return new_dag
-
-    def fix_dag(dag: list, start_nodes: set, output_node=None):
-        dag.sort(key=lambda x: (x[0], x[1]))
-        new_dag = fix_dag_one_way(dag, start_nodes, True)
-
-        if output_node:
-            new_dag = fix_dag_one_way(new_dag, {output_node}, False)
-
-        new_dag.sort(key=lambda x: (x[0], x[1]))
-
-        if not new_dag:
-            return [(max(start_nodes), output_node, shared_cnn.identity.__name__)]
-
-        return new_dag
-
-    def sample_fixed_dag(dag_probs, all_connections):
-        dag = sample_dags(dag_probs, all_connections)
-        dag = fix_dag(dag, {0, 1}, 1 + args.num_blocks)
-        return dag
-
-    def sample_fixed_dags(dags_probs, all_connections):
-        return [sample_fixed_dag(dag_probs, all_connections=all_connections) for dag_probs in dags_probs]
-
-    def get_random():
-        random_dag = []
-        # random_dag.append((0, 1, np.random.choice(shared_cnn.CNNCell.default_layer_types)))
-        for i in range(2, 2+args.num_blocks):
-            # previous_layers = np.random.choice(possible_connections, 2, replace=True)
-            possible_connections = list(connection for connection in all_connections if connection[1] == i)
-            # print(possible_connections)
-            ids = np.random.choice(len(possible_connections), 2, replace=True)
-            for id in ids:
-                random_dag.append(possible_connections[id])
-
-        return trim_dag(random_dag, {0, 1}, 1 + args.num_blocks)
 
     num_batches = 1
     total_loss = 0
@@ -299,19 +178,18 @@ def main():
                             total_loss = 0
                             num_batches = 0
 
-                            sT = start()
-                            dags = sample_fixed_dags(dags_probs, all_connections)
+                            with Timer("to_cuda", logger.info):
+                                dags = sample_fixed_dags(dags_probs, all_connections, args.num_blocks)
 
-                            logger.info(dags)
+                                logger.info(dags)
 
-                            cnn_optimizer.full_reset_grad()
-                            cnn.to_gpu(dags)
-                            cnn_optimizer.to(cnn.get_parameters(last_dags), cpu)
-                            cnn_optimizer.to(cnn.get_parameters(dags), gpu)
+                                cnn_optimizer.full_reset_grad()
+                                cnn.to_gpu(dags)
+                                cnn_optimizer.to(cnn.get_parameters(last_dags), cpu)
+                                cnn_optimizer.to(cnn.get_parameters(dags), gpu)
 
-                            last_dags = dags
-                            dropout_opt.zero_grad()
-                            stop(sT, "to_cuda", logger)
+                                last_dags = dags
+                                dropout_opt.zero_grad()
                             get_new_network = False
                             num_dropout_batches_items = 0
 
@@ -482,16 +360,15 @@ def main():
                                 total_acc = 0
                                 total_loss = 0
 
-                                sT = start()
+                                with Timer("to_cuda", logger.info):
+                                    dags = sample_fixed_dags(dags_probs, all_connections, args.num_blocks)
 
-                                dags = sample_fixed_dags(dags_probs, all_connections)
+                                    logger.info(dags)
 
-                                logger.info(dags)
+                                    cnn.to_gpu(dags)
 
-                                cnn.to_gpu(dags)
+                                    last_dags = dags
 
-                                last_dags = dags
-                                stop(sT, "to_cuda", logger)
                                 get_new_network = False
                                 num_dropout_batches_items = 0
